+++ packages/varmint/__tests__/filebox.test.ts
@@ -75,78 +75,4 @@ describe(`Filebox`, () => {
 		expect(result).toBe(`The best way to predict the future is to invent it.`)
 		expect(utils.put).toHaveBeenCalledTimes(0)
 	})
-	test(`mode:read (cache-miss)`, async () => {
-		const info: string[] = []
-		const squirrel = new Squirrel(`read`, tempDir.name)
-		fs.mkdirSync(path.join(tempDir.name, `hello`))
-		fs.writeFileSync(
-			path.join(tempDir.name, `hello/casa.input.json`),
-			`[\n\t"http://localhost:12500"\n]`,
-		)
-		const fetcher = squirrel.add(`hello`, async (url: string) => {
-			return fetch(url).then((response) => response.text())
-		})
-		let caught: Error | undefined
-		try {
-			await fetcher.for(`home`).get(`http://localhost:12500`)
-		} catch (thrown) {
-			if (thrown instanceof Error) {
-				console.error(`💥`, thrown)
-				caught = thrown
-			}
-		} finally {
-			expect(caught).toBeInstanceOf(Error)
-			expect(utils.put).toHaveBeenCalledTimes(0)
-		}
-	})
-	test(`flushing untouched files`, async () => {
-		const setup = spawn(
-			`node`,
-			[`--experimental-strip-types`, `global-setup.node.ts`, tempDir.name],
-			{
-				stdio: `inherit`,
-				cwd: path.join(import.meta.dirname, `isolation`),
-			},
-		)
-		await new Promise((resolve) => setup.on(`exit`, resolve))
-		fs.mkdirSync(path.join(tempDir.name, `other`))
-		console.log(`tempDir contents:`, fs.readdirSync(tempDir.name))
-		fs.writeFileSync(
-			path.join(tempDir.name, `rand`, `some-random-file.whatever`),
-			`{}`,
-		)
-		fs.writeFileSync(
-			path.join(tempDir.name, `myStreamer`, `another-random-file.whatever`),
-			`{}`,
-		)
-		expect(fs.readdirSync(tempDir.name)).toEqual([`myStreamer`, `other`, `rand`])
-		expect(fs.readdirSync(path.join(tempDir.name, `rand`))).toEqual([
-			`my-rand.input.json`,
-			`my-rand.output.json`,
-			`some-random-file.whatever`,
-		])
-		expect(fs.readdirSync(path.join(tempDir.name, `myStreamer`))).toEqual([
-			`another-random-file.whatever`,
-			`myAsyncIterable.input.json`,
-			`myAsyncIterable.stream.txt`,
-		])
-		const teardown = spawn(
-			`node`,
-			[`--experimental-strip-types`, `global-teardown.node.ts`],
-			{
-				stdio: `inherit`,
-				cwd: path.join(import.meta.dirname, `isolation`),
-			},
-		)
-		await new Promise((resolve) => teardown.on(`exit`, resolve))
-		expect(fs.readdirSync(tempDir.name)).toEqual([`myStreamer`, `rand`])
-		expect(fs.readdirSync(path.join(tempDir.name, `rand`))).toEqual([
-			`my-rand.input.json`,
-			`my-rand.output.json`,
-		])
-		expect(fs.readdirSync(path.join(tempDir.name, `myStreamer`))).toEqual([
-			`myAsyncIterable.input.json`,
-			`myAsyncIterable.stream.txt`,
-		])
-	})
 })
